requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
aiohttp==3.10.5
sentence-transformers==3.1.1
numpy==1.26.4
scipy==1.14.1
apscheduler==3.10.4
python-dotenv==1.0.1
python-multipart==0.0.12
jinja2==3.1.4
